requirements.txt
rpyc>=5.3.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0
sqlalchemy>=2.0.0
aiohttp>=3.8.0
websockets>=11.0
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2023.3
